Separates sparkles-icon and renames file
